requirements.txt
-e git://github.com/stephenmcd/mezzanine.git#egg=mezzanine
django-compressor
django-extensions
django-celery
django-mailbox
feedparser
gunicorn
psycopg2
python-memcached
